Move new components to correct folders
